# microservice/source/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Inventory API
  description: |-
    A demo inventory management system.
  version: 2.0.0
servers:
  - url: https://inventory-api.igraphql.co/api/v2
    description: Microservice
...
paths:
  /assets/{skuId}:
    post:
      tags:
        - Assets
      description: Creates a product's assets.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssets'
      responses:
        "201":
          description: Asset(s) successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode201"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: SKU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "409":
          description: Asset(s) already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode409"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    get:
      tags:
        - Assets
      description: Retrieves a product's assets.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - $ref: '#/components/parameters/AssetID'
        - $ref: '#/components/parameters/Field'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Tag'
        - $ref: '#/components/parameters/URL'
      responses:
        "200":
          description: Asset(s) successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssets"
        "404":
          description: Asset(s) or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    put:
      tags:
        - Assets
      description: Updates a product's assets.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssets'
      responses:
        "200":
          description: Assets successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: Asset(s) or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Assets
      description: Deletes a product's assets.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      responses:
        "200":
          description: Assets successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: SKU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
  /assets/{skuId}/{assetId}:
    get:
      tags:
        - Assets
      description: Retrieves a product asset.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: assetId
          in: path
          description: Asset ID.
          required: true
          schema:
            type: integer
            example: 14
        - $ref: '#/components/parameters/Field'
      responses:
        "200":
          description: Asset successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAsset"
        "404":
          description: Asset or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    put:
      tags:
        - Assets
      description: Updates a product asset.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: assetId
          in: path
          description: Asset ID.
          required: true
          schema:
            type: integer
            example: 14
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAsset'
      responses:
        "200":
          description: Asset successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: Asset or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Assets
      description: Deletes a product asset.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: assetId
          in: path
          description: Asset ID.
          required: true
          schema:
            type: integer
            example: 14
      responses:
        "200":
          description: Asset successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: Asset or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
...
components:
  parameters:
...
    AssetID:
      name: asset_id
      in: query
      description: Asset ID.
      schema:
        type: integer
...
    Field:
      name: field
      in: query
      description: Limit the fields returned by the query. Can be specified multiple times.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
...
    Limit:
      name: limit
      in: query
      description: Limit the number of values returned by the query.
      schema:
        type: integer
...
    Offset:
      name: offset
      in: query
      description: Offset the number of values returned by the query.
      schema:
        type: integer
...
    Tag:
      name: tag
      in: query
      description: Asset tag.
      schema:
        type: string
...
    URL:
      name: url
      in: query
      description: Asset URL
      schema:
        type: string
...
  schemas:
    CreateAssets:
      type: array
      items:
        type: object
        required:
          - tag
          - url
        properties:
          tag:
            type: string
            example: thumbnail
          url:
            type: string
            example: /images/assets/96070.png
    GetAsset:
      type: object
      properties:
        asset_id:
          type: integer
          example: 14
        created:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        sku_id:
          type: integer
          example: 96070
        tag:
          type: string
          example: thumbnail
        updated:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        url:
          type: string
          example: /images/assets/96070.png
    GetAssets:
      type: array
      items:
        type: object
        properties:
          asset_id:
            type: integer
            example: 14
          created:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          sku_id:
            type: integer
            example: 96070
          tag:
            type: string
            example: thumbnail
          updated:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          url:
            type: string
            example: /images/assets/96070.png
    UpdateAsset:
      type: object
      properties:
        tag:
          type: string
          example: thumbnail
        url:
          type: string
          example: /images/assets/96070.png
    UpdateAssets:
      type: array
      items:
        type: object
        required:
          - asset_id
        properties:
          asset_id:
            type: integer
            example: 14
          tag:
            type: string
            example: thumbnail
          url:
            type: string
            example: /images/assets/96070.png
...
    HttpCode200:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: ok
    HttpCode201:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: created
    HttpCode400:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: bad request
    HttpCode404:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: not found
    HttpCode409:
      type: object
      properties:
        status:
          type: integer
          example: 409
        message:
          type: string
          example: conflict
    HttpCode415:
      type: object
      properties:
        status:
          type: integer
          example: 415
        message:
          type: string
          example: unsupported media type
    HttpCode500:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: internal server error